move UMAP and IPC version of graph demo into demos/ipc folder + cleanup
