+++ .gitignore
@@ -23,7 +23,6 @@ modules/*/include/napi-inl.h
 modules/*/include/napi-inl.deprecated.h
 modules/*/include/node_api.h
 
-modules/demo/umap/data
-modules/demo/graph/data
+modules/demo/**/*/data
 
 modules/**/.eslintrc.js
