+++ lerna.json
@@ -7,10 +7,10 @@
     "modules/*",
     "modules/demo/graph",
     "modules/demo/luma",
-    "modules/demo/umap",
     "modules/demo/xterm",
     "modules/demo/cugraph",
     "modules/demo/deck/*",
+    "modules/demo/ipc/*",
     "modules/demo/tfjs/*"
   ]
 }
