+++ modules/cudf/package.json
@@ -30,7 +30,7 @@
     "apache-arrow": "^2.0.0"
   },
   "devDependencies": {
-    "ix": "^4.1.0"
+    "ix": "^4.3.0"
   },
   "files": [
     "LICENSE",
