--- modules/demo/graph/.env
@@ -1,3 +1,0 @@
-EXTEND_ESLINT=true
-SKIP_PREFLIGHT_CHECK=true
-REACT_APP_ENVIRONMENT=browser
