--- modules/demo/graph/.eslintignore
@@ -1,1 +1,0 @@
-*
