--- modules/demo/graph/public/video.html
@@ -1,69 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8" />
-    <title>Graph Demo</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="theme-color" content="#2e2e2e" />
-</head>
-<body style="background:#2e2e2e; margin:0;">
-    <video autoplay muted width="800" height="600"></video>
-    <script src="https://cdn.jsdelivr.net/npm/simple-peer@9.7.2/simplepeer.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/2.3.0/socket.io.js"></script>
-    <script>
-        var video = document.querySelector('video');
-        var peer = new SimplePeer({
-            initiator: true, trickle: false,
-            // Remove bandwidth restrictions
-            // https://github.com/webrtc/samples/blob/89f17a83ed299ef28d45b933419d809b93d41759/src/content/peerconnection/bandwidth/js/main.js#L240
-            sdpTransform: (sdp) => sdp.replace(/b=AS:.*\r\n/, '').replace(/b=TIAS:.*\r\n/, '')
-        });
-        var sock = io({ transports: ['websocket'], reconnection: false });
-
-        sock.on('signal', (data) => peer.signal(data));
-        peer.on('signal', (data) => sock.emit('signal', data));
-
-        peer.on('stream', (stream) => {
-
-            ('srcObject' in video)
-                ? (video.srcObject = stream)
-                : (video.src = window.URL.createObjectURL(stream));  // for older browsers
-            video.play();
-
-            peer.send('play');
-        });
-
-        dispatchRemoteEvent(video, 'blur');
-        dispatchRemoteEvent(video, 'focus');
-        dispatchRemoteEvent(video, 'wheel');
-        // dispatchRemoteEvent(window, 'beforeunload');
-        dispatchRemoteEvent(document, 'keyup');
-        dispatchRemoteEvent(document, 'keydown');
-        dispatchRemoteEvent(document, 'keypress');
-        dispatchRemoteEvent(video, 'mouseup');
-        dispatchRemoteEvent(video, 'mousemove');
-        dispatchRemoteEvent(video, 'mousedown');
-        dispatchRemoteEvent(video, 'mouseenter');
-        dispatchRemoteEvent(video, 'mouseleave');
-
-        function dispatchRemoteEvent(target, type) {
-            target.addEventListener(type, (e) => {
-                peer.send(JSON.stringify({
-                    event: Object
-                        .getOwnPropertyNames(Object.getPrototypeOf(e))
-                        .reduce((event, field) => {
-                            switch (typeof e[field]) {
-                                case 'object': break;
-                                case 'symbol': break;
-                                case 'function': break;
-                                default: event[field] = e[field];
-                            }
-                            return event;
-                        }, { type: e.type })
-                }));
-            });
-        };
-
-    </script>
-</body>
-</html>
