+++ modules/demo/ipc/graph/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "graph-demo",
+  "name": "ipc-graph-demo",
   "version": "0.0.0",
   "private": true,
   "dependencies": {
@@ -13,7 +13,8 @@
     "@nvidia/glfw": "0.0.1",
     "ix": "^4.3.0",
     "react": "^16.13.1",
-    "react-dom": "^16.13.1"
+    "react-dom": "^16.13.1",
+    "zeromq": "6.0.0-beta.6"
   },
   "scripts": {
     "start": "node ./index.js"
