+++ modules/demo/ipc/graph/python/test_data.py
@@ -16,17 +16,16 @@
 import cudf
 import cugraph
 import numpy as np
+import pandas as pd
+import datetime as dt
 from .convert_matrix import from_cudf_edgelist
 from .graph_components import (
     annotate_nodes,
-    annotate_edges,
-    category_to_color
+    annotate_edges
 )
 
 
 def make_synthetic_dataset(**kwargs):
-    import pandas as pd
-    import datetime as dt
     kwargs.update(direct=True)
     df = cudf.DataFrame.from_pandas(pd.DataFrame({
         "src": [0, 1, 2, 3],
@@ -78,42 +77,4 @@ def make_and_shape_hypergraph(df, **kwargs):
     # Add vis components 
     nodes = annotate_nodes(graph, nodes, edges)
     edges = annotate_edges(graph, nodes, edges)
-    return graph, nodes, edges
-
-
-def make_capwin_dataset(**kwargs):
-
-    def drop_index(df):
-        return df.reset_index(drop=True)
-
-    def smoosh(df):
-        size = sum([df[x].dtype.itemsize for x in df])
-        data = drop_index(drop_index(df).stack()).data
-        dtype = cudf.utils.dtypes.min_unsigned_type(0, size*8)
-        return cudf.core.column.NumericalColumn(data, dtype=dtype)
-
-    def add_edge_colors(edges, category):
-        colors = drop_index(category_to_color(edges[category], color_palette=[
-            #  ADDRESS   AUTH KEYS CREDENTIALS       EMAIL      FALSE
-            4294967091, 4294410687, 4293138972, 4281827000,  33554431
-        ]).astype(np.uint32))
-        return edges.assign(color=smoosh(cudf.DataFrame({
-            "src": drop_index(colors), "dst": drop_index(colors)
-        })).astype(np.uint64), src_color=colors)
-    
-    df = cudf.read_csv("data/pii_sample_for_viz.csv")
-    df = df[["src_ip", "dest_ip", "pii", "timestamp"]]
-    df["timestamp"] = cudf.to_datetime(df["timestamp"], format="%m/%d/%y %H:%M")
-    # Create graph
-    graph, nodes, edges = from_cudf_edgelist(df, "src_ip", "dest_ip")
-    # Add vis components 
-    nodes = nodes.rename({"node": "name"}, axis=1, copy=False)
-    nodes = annotate_nodes(graph, nodes, edges)
-    # add edge colors
-    edges = add_edge_colors(edges, "pii")
-    print(edges.query("src_color != 33554431")["src"].value_counts())
-    print(edges.query("src_color != 33554431")["dst"].value_counts())
-    # add edge names
-    edges["name"] = edges["src_ip"] + " -> " + edges["dest_ip"] + \
-        ("\nPII: " + edges["pii"]).replace("\nPII: FALSE", "")
     return graph, nodes, edges
