+++ modules/demo/ipc/graph/src/loader.js
