+++ modules/demo/ipc/umap/layers/edge/bezier-curve-layer-fragment.glsl.js
