+++ modules/demo/ipc/umap/layers/edge/bezier-curve-layer.js
