+++ modules/demo/ipc/umap/layers/node/node-layer-fragment.glsl.js
