+++ scripts/demo/linux.sh
@@ -29,8 +29,8 @@ fi
 
 if [[ "$DEMO" == "" ]]; then
     DEMOS="
-    $(echo modules/demo/{luma,umap,graph,xterm}/package.json)
-    $(find modules/demo/{deck,tfjs} -maxdepth 2 -type f -name 'package.json')
+    $(echo modules/demo/{luma,graph,xterm}/package.json)
+    $(find modules/demo/{deck,tfjs,ipc} -maxdepth 2 -type f -name 'package.json')
     ";
     DEMOS="$(echo -e "$DEMOS" | grep -v node_modules | sort -Vr)";
     DEMOS=(${DEMOS});
@@ -52,7 +52,7 @@ fi
 ARGS="${@:-}";
 
 if [[ "$DEMO" =~ "modules/demo/luma" ]]; then ARGS="${@:-01}";
-elif [[ "$DEMO" =~ "modules/demo/umap" ]]; then ARGS="${@:-tcp://0.0.0.0:6000}";
+elif [[ "$DEMO" =~ "modules/demo/ipc/umap" ]]; then ARGS="${@:-tcp://0.0.0.0:6000}";
 fi
 
 NODE_ENV=production exec node -r esm --trace-uncaught "$DEMO" $ARGS
